Refactor Duke to Coinflip
