+++ build.gradle
@@ -29,11 +29,11 @@ test {
 }
 
 application {
-    mainClass.set("seedu.duke.Duke")
+    mainClass.set("seedu.coinflip.Coinflip")
 }
 
 shadowJar {
-    archiveBaseName.set("duke")
+    archiveBaseName.set("coinflip")
     archiveClassifier.set("")
 }
