+++ src/test/java/seedu/coinflip/CoinflipTest.java
@@ -1,10 +1,10 @@
-package seedu.duke;
+package seedu.coinflip;
 
 import static org.junit.jupiter.api.Assertions.assertTrue;
 
 import org.junit.jupiter.api.Test;
 
-class DukeTest {
+class CoinflipTest {
     @Test
     public void sampleTest() {
         assertTrue(true);
